Invincibility frames, knockback, and more
Main stuff was invincibility frames and knockback implementations.
Knockback also uses the screen shake script, which is pretty neat by the
way. The knockback causes some issues which I will elaborate on down
below.

Invincibility frames can be applied to any damageable object. This can
help prevent spamming weak notes at a zombie over and over again.

When the player is hit by the zombie, he will be invincible for 50
frames (can be changed via player health script). Knockback will not
happen until the player is out of the invincibility frames. (There is no
effect for being invincible yet.)

Projectiles can now have a knockback property passed to it (force for
both the x and y directions). The gravity property may also be set now.

---
Here's the problem with knockback:
The way it works is that it will apply a force to the damageable object.
This doesn't work so well with the character controller script, as it
will hard reset the horizontal movement, thus ending the force added for
knockback very abruptly. Also, when getting knocked back into a slope,
the player is shot vertically at some fairly impressive speeds. I'm not
sure how to fix either of these problems, and I probably won't until
daytime because it's 5:15 AM here and I want to go to sleep.